Add unicum/group as additional_parents
Some categories contain purely unique products (e.g. "Person"), some others contain products that each have multiple instances (e.g. "Camera"). In the latter case, the rendered list is, more specifically, not actually a "list of cameras", but a "list of camera types".

Unfortunately, sometimes a category is child of Unicum *and* another category that is Group.
In the dataset at hand, these are:
- Parent unique, Child group:
  - Activity, Game
  - TopicalConcept, Type
- Parent group, child unique:
  - Device, Supercomputer
  - MeanOfTransportation, SpaceStation
  - MusicalWork, ClassicalMusicComposition
  - MusicalWork, Musical
  - MusicalWork, NationalAnthem
  - MusicalWork, Opera
  - MusicalWork, Song
Not 100% sure this is the best way to proceed, because since categories inherit their parent categories' properties, the above listed categories are both at once. E.g. a "Game" is both a group and unicum (there are millions of copies of "Monopoly", but Monopoly is an unambiguous, unique Activity). But as long as belonging to unicum/group is of no consequences for a category at all, it can stay like that.

Instead of the additional_parents approach it would be better to just list all categories in an array with mutliple parents listed (like it is in the database later on). But the tree structure is far more intuitive for the moment